LPS-171146 MultiLanguageConditionEvaluator: SF, inline
