+++ modules/apps/portal-workflow/portal-workflow-kaleo-runtime-impl/src/main/java/com/liferay/portal/workflow/kaleo/runtime/internal/node/ConditionNodeExecutor.java
@@ -14,8 +14,16 @@
 
 package com.liferay.portal.workflow.kaleo.runtime.internal.node;
 
+import com.liferay.osgi.service.tracker.collections.map.ServiceTrackerMap;
+import com.liferay.osgi.service.tracker.collections.map.ServiceTrackerMapFactory;
+import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.exception.PortalException;
+import com.liferay.portal.kernel.util.ClassUtil;
+import com.liferay.portal.kernel.util.GetterUtil;
+import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.workflow.kaleo.definition.NodeType;
+import com.liferay.portal.workflow.kaleo.definition.ScriptLanguage;
+import com.liferay.portal.workflow.kaleo.definition.exception.KaleoDefinitionValidationException;
 import com.liferay.portal.workflow.kaleo.model.KaleoCondition;
 import com.liferay.portal.workflow.kaleo.model.KaleoInstanceToken;
 import com.liferay.portal.workflow.kaleo.model.KaleoNode;
@@ -30,7 +38,10 @@ import com.liferay.portal.workflow.kaleo.service.KaleoInstanceLocalService;
 
 import java.util.List;
 
+import org.osgi.framework.BundleContext;
+import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Deactivate;
 import org.osgi.service.component.annotations.Reference;
 
 /**
@@ -42,6 +53,46 @@ public class ConditionNodeExecutor extends BaseNodeExecutor {
 	@Override
 	public NodeType getNodeType() {
 		return NodeType.CONDITION;
+	}
+
+	@Activate
+	protected void activate(BundleContext bundleContext) {
+		_serviceTrackerMap = ServiceTrackerMapFactory.openSingleValueMap(
+			bundleContext, ConditionEvaluator.class, "(scripting.language=*)",
+			(serviceReference, emitter) -> {
+				Object propertyValue = serviceReference.getProperty(
+					"scripting.language");
+
+				ConditionEvaluator conditionEvaluator =
+					bundleContext.getService(serviceReference);
+
+				try {
+					for (String scriptingLanguage :
+							GetterUtil.getStringValues(
+								propertyValue,
+								new String[] {String.valueOf(propertyValue)})) {
+
+						emitter.emit(
+							_getConditionEvaluatorKey(
+								scriptingLanguage,
+								ClassUtil.getClassName(conditionEvaluator)));
+					}
+				}
+				catch (KaleoDefinitionValidationException
+							kaleoDefinitionValidationException) {
+
+					throw new RuntimeException(
+						kaleoDefinitionValidationException);
+				}
+				finally {
+					bundleContext.ungetService(serviceReference);
+				}
+			});
+	}
+
+	@Deactivate
+	protected void deactivate() {
+		_serviceTrackerMap.close();
 	}
 
 	@Override
@@ -64,8 +115,7 @@ public class ConditionNodeExecutor extends BaseNodeExecutor {
 			_kaleoConditionLocalService.getKaleoNodeKaleoCondition(
 				currentKaleoNode.getKaleoNodeId());
 
-		String transitionName = _conditionEvaluator.evaluate(
-			kaleoCondition, executionContext);
+		String transitionName = _evaluate(kaleoCondition, executionContext);
 
 		_kaleoInstanceLocalService.updateKaleoInstance(
 			kaleoInstanceToken.getKaleoInstanceId(),
@@ -92,13 +142,45 @@ public class ConditionNodeExecutor extends BaseNodeExecutor {
 		List<PathElement> remainingPathElements) {
 	}
 
-	@Reference(target = "(!(scripting.language=*))")
-	private ConditionEvaluator _conditionEvaluator;
+	private String _evaluate(
+			KaleoCondition kaleoCondition, ExecutionContext executionContext)
+		throws PortalException {
+
+		String conditionEvaluatorKey = _getConditionEvaluatorKey(
+			kaleoCondition.getScriptLanguage(),
+			StringUtil.trim(kaleoCondition.getScript()));
+
+		ConditionEvaluator conditionEvaluator = _serviceTrackerMap.getService(
+			conditionEvaluatorKey);
+
+		if (conditionEvaluator == null) {
+			throw new IllegalArgumentException(
+				"No condition evaluator found for script language " +
+					conditionEvaluatorKey);
+		}
+
+		return conditionEvaluator.evaluate(kaleoCondition, executionContext);
+	}
+
+	private String _getConditionEvaluatorKey(
+			String language, String conditionEvaluatorClassName)
+		throws KaleoDefinitionValidationException {
+
+		ScriptLanguage scriptLanguage = ScriptLanguage.parse(language);
+
+		if (scriptLanguage.equals(ScriptLanguage.JAVA)) {
+			return language + StringPool.COLON + conditionEvaluatorClassName;
+		}
+
+		return language;
+	}
 
 	@Reference
 	private KaleoConditionLocalService _kaleoConditionLocalService;
 
 	@Reference
 	private KaleoInstanceLocalService _kaleoInstanceLocalService;
+
+	private ServiceTrackerMap<String, ConditionEvaluator> _serviceTrackerMap;
 
 }
